requirements.txt
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
anthropic>=0.25.0
google-generativeai>=0.7.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
aiofiles>=23.2.1
python-multipart>=0.0.6
numpy>=1.24.0
